services/python/pyproject.toml
[project]
name = "ai-service"
authors = [
    {name = "Abhishek", email = "abhishek@pipeshub.com"}
]
description = "AI Pipeline Service"
version = "0.1"
readme = "README.md"
requires-python = ">=3.7"
dependencies = [
    "pydantic<3.0.0",
    "numpy<2",
    "fastapi==0.115.6",
    "uvicorn==0.30.6",
    "python-dotenv==1.0.1",
    "pandas==2.2.3",
    "openpyxl==3.1.5",
    "python-multipart==0.0.18",
    "jsonschema==4.23.0",
    "nats-py==2.1.0",
    "tenacity==8.5.0",
    "azure-ai-formrecognizer==3.3.3",
    "requests-mock==1.12.1",
    "protobuf==3.20.3",
    "aiohttp==3.11.2",
    "python-jose==3.4.0",
    "requests-toolbelt==1.0.0",
    "spacy<=3.7.5",
    "python-docx==1.1.2",
    "uuid==1.30",
    "confluent-kafka==2.8.0",
    "dependency-injector==4.44.0",
    "pdf2image==1.17.0",
    "PyMuPDF==1.24.14",
    "ocrmypdf==16.8.0",
    "langchain==0.3.19",
    "qdrant-client==1.13.1",
    "langchain-experimental==0.3.4",
    "sentence-transformers==3.4.1",
    "langchain-huggingface==0.1.2",
    "langchain-qdrant==0.2.0",
    "fastembed==0.5.1",
    "spacy<=3.7.5",
    "nltk==3.9.1",
    "langchain-openai==0.3.6",
    "langchain-anthropic==0.3.8",
    "docling==2.25.1",
    "etcd3==0.12.0",
    "python-arango==8.1.5",
    "redis==5.2.1",
    "docx2python==3.5.0",
    "google-api-python-client==2.161.0",
    "aiolimiter==1.2.1",
    "google-auth-oauthlib==1.2.1",
    "celery==5.4.0",
    "requests==2.32.3",
    "langchain-community==0.3.18",
    "Jinja2==3.1.6",
    "cryptography==44.0.1",
    "sentence-transformers==3.4.1"
]

[tool.setuptools.packages.find]
exclude = ["modules"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project.scripts]
module = "app.indexing_main:run"
